Refactor DialogBox.java to gui package
